% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{calculate_C}
\alias{calculate_C}
\title{Function to simulate air to air convection via heat convection equation. This is done in 3D.
In this function, air temperature is updated and ready to use for the sensible heat flux calculation.}
\usage{
calculate_C(temp_air)
}
\arguments{
\item{temp_air}{Air temperature}
}
\value{
total_convection Convection flux per voxel
}
\description{
Function to simulate air to air convection via heat convection equation. This is done in 3D.
In this function, air temperature is updated and ready to use for the sensible heat flux calculation.
}
