% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{calculate_D}
\alias{calculate_D}
\title{Function to simulate air to air diffusion via Fourier's law of heat conduction. This is done in 3D.
In this function, air temperature is updated and ready to use for the sensible heat flux calculation.}
\usage{
calculate_D(temp_air, temp_soil)
}
\arguments{
\item{temp_air}{Air temperature}

\item{temp_soil}{Ground surface temperature}
}
\value{
total_diffusion Diffusion flux per voxel
}
\description{
Function to simulate air to air diffusion via Fourier's law of heat conduction. This is done in 3D.
In this function, air temperature is updated and ready to use for the sensible heat flux calculation.
}
