% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{calculate_G}
\alias{calculate_G}
\title{Function to calculate ground heat flux G}
\usage{
calculate_G(net_rad_ground)
}
\arguments{
\item{net_rad_ground}{Net radiation at ground surface}
}
\value{
G Ground heat flux
}
\description{
Function to calculate ground heat flux G
}
