% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{calculate_H}
\alias{calculate_H}
\title{Function to calculate sensible heat transfer between surface and air}
\usage{
calculate_H(T_surf)
}
\arguments{
\item{T_surf}{Surface temperature}
}
\value{
total_convection Convection flux per voxel
}
\description{
Function to calculate sensible heat transfer between surface and air
}
