% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{calculate_LE}
\alias{calculate_LE}
\title{Function to calculate latent heat flux LE (~ evapotranspiration, ET: LE = L_v . ET with L_v = latent heat of vaporization of water (J/kg))}
\usage{
calculate_LE(temperature, net_rad, G)
}
\arguments{
\item{temperature}{Forest surface temperature}

\item{net_rad}{Net radiation}
}
\value{
LE Latent heat flux
}
\description{
Function to calculate latent heat flux LE (~ evapotranspiration, ET: LE = L_v . ET with L_v = latent heat of vaporization of water (J/kg))
}
