% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_temperatures.R
\name{core_temp}
\alias{core_temp}
\title{Function to define the initial forest core temperature based on macro temperature, season and time of day}
\usage{
core_temp(T_macro)
}
\arguments{
\item{T_macro}{Macrotemperature}
}
\value{
core_temp Forest core temperature
}
\description{
Function to define the initial forest core temperature based on macro temperature, season and time of day
}
