% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input_variables_and_params.R
\name{create_input_data}
\alias{create_input_data}
\title{Generate standard input data for ForEdgeClim}
\usage{
create_input_data(datetime)
}
\arguments{
\item{datetime}{POSIXct object}
}
\value{
global variables and parameters
}
\description{
Generate standard input data for ForEdgeClim
}
