% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_drivers_constants_parameters.R
\name{create_input_drivers}
\alias{create_input_drivers}
\title{Generate input driver variables}
\usage{
create_input_drivers()
}
\arguments{
\item{datetime}{POSIXct object}
}
\value{
input driver variables
}
\description{
Generate input driver variables
}
