% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_drivers_constants_parameters.R
\name{create_physical_constants}
\alias{create_physical_constants}
\title{Generate (physical) constants}
\usage{
create_physical_constants()
}
\value{
(physical) constants
}
\description{
Generate (physical) constants
}
