% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{de_dT}
\alias{de_dT}
\title{Function to calculate the derivative of the energy balance closure error to surface temperature}
\usage{
de_dT(temp, net_rad)
}
\arguments{
\item{temp}{Forest surface temperature}

\item{net_rad}{Net radiation}
}
\value{
Derivative of energy balance wrt forest surface temperature
}
\description{
Function to calculate the derivative of the energy balance closure error to surface temperature
}
