% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid_and_DTM_from_Vox.R
\name{generate_DTM_grid_Vox}
\alias{generate_DTM_grid_Vox}
\title{Generate a voxelised point density grid from a vox file}
\usage{
generate_DTM_grid_Vox(vox_file)
}
\arguments{
\item{vox_file}{The vox file}
}
\value{
A data.table with columns: X, Y, Z, and density.
}
\description{
This function reads a vox file and creates a 3D voxel grid where each voxel represents
a 1x1x1m^3 volume. The density of each voxel is computed using Plant Area Density (PAD) from AMAPVox.
The output variable in AMAPVox representing PAD is 'Attenuation unbiased maximum likelihood estimator based
on free path length (FPL MLE, F. Pimont)'. It takes into account beam attenuation and corrects for
the fact that not all beams can 'see' everything (occlustion).
The output is a data.table containing rescaled X, Y, and Z coordinates and a normalized density value.
}
\details{
Additionally, a DTM is generated.
}
