% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid_from_fiction.R
\name{generate_fictive_grid}
\alias{generate_fictive_grid}
\title{Generate a virtual structural grid with trees}
\usage{
generate_fictive_grid(season)
}
\arguments{
\item{season}{Season as string: "summer", "winter", "spring", "autumn"}
}
\value{
A dataframe with the generated voxel structure
}
\description{
Generate a virtual structural grid with trees
}
