% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_temperatures.R
\name{get_season}
\alias{get_season}
\title{Function to determine season}
\usage{
get_season(datetime)
}
\arguments{
\item{datetime}{Datetime object representing the current simulation time}
}
\value{
'winter' or 'spring' or 'summer' or 'autumn'
}
\description{
Function to determine season
}
