% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_DTS_observations}
\alias{import_DTS_observations}
\title{Function to import hourly DTS data observations}
\usage{
import_DTS_observations()
}
\value{
Filtered CSV file
}
\description{
Function to import hourly DTS data observations
}
