% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_PAR_observations}
\alias{import_PAR_observations}
\title{Function to import Odyssey PAR observations (PAR radiation)}
\usage{
import_PAR_observations()
}
\value{
PAR radiation
}
\description{
Function to import Odyssey PAR observations (PAR radiation)
}
