% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_PE_observations}
\alias{import_PE_observations}
\title{Function to import macro temperature from Plant Ecology lab}
\usage{
import_PE_observations()
}
\value{
macro temperature and downward longwave radiation
}
\description{
Function to import macro temperature from Plant Ecology lab
}
