% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_pyr_observations}
\alias{import_pyr_observations}
\title{Function to import Delta-T pyranometer observations (direct and diffuse light)}
\usage{
import_pyr_observations()
}
\value{
Direct solar beam and diffuse radiation
}
\description{
Function to import Delta-T pyranometer observations (direct and diffuse light)
}
