% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_RMI_observations}
\alias{import_RMI_observations}
\title{Function to import RMI national weather station observation (macro temperature)}
\usage{
import_RMI_observations()
}
\value{
macro temperature and downward longwave radiation
}
\description{
Function to import RMI national weather station observation (macro temperature)
}
