% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_temperatures.R
\name{is_night_or_day}
\alias{is_night_or_day}
\title{Function to determine night or day}
\usage{
is_night_or_day(datetime)
}
\arguments{
\item{datetime}{Datetime object representing the current simulation time}
}
\value{
'night' or 'day'
}
\description{
Function to determine night or day
}
