% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM_LW.R
\name{longwave_two_stream_RTM}
\alias{longwave_two_stream_RTM}
\title{2D Longwave Two-Stream Radiative Transfer Model (RTM)}
\usage{
longwave_two_stream_RTM(
  voxel_grid,
  micro_grid,
  lw_two_stream,
  F_sky_lw,
  omega_g_lw_v,
  omega_g_lw_h,
  macro_temp,
  Kd_lw_v,
  Kd_lw_h,
  omega_lw,
  beta_lw,
  season,
  N_or_D
)
}
\arguments{
\item{voxel_grid}{Dataframe containing voxel density data (columns: X, Y, Z, density).}

\item{micro_grid}{Dataframe containing microclimate temperature data (columns: x, y, z, temperature, T_soil).}

\item{lw_two_stream}{Function that calculates longwave two-stream radiation for a single column.}

\item{F_sky_lw}{Longwave sky flux.}

\item{omega_g_lw_v}{Ground reflectance for vertical LW RTM.}

\item{omega_g_lw_h}{Ground reflectance for horizontal LW RTM.}

\item{macro_temp}{Macroclimate temperature outside the forest.}

\item{Kd_lw_v}{Diffuse extinction coefficient for vertical LW RTM.}

\item{Kd_lw_h}{Diffuse extinction coefficient for horizontal LW RTM.}

\item{omega_lw}{Single-scattering albedo for longwave RTM.}

\item{beta_lw}{Phase function parameter for longwave RTM.}

\item{season}{Season (e.g., "summer", "winter") used in the RTM.}

\item{N_or_D}{Indicator for night or day.}
}
\value{
A dataframe with computed longwave fluxes and net longwave radiation.
}
\description{
This function performs a two-directional (vertical & horizontal) longwave
two-stream RTM to estimate radiation fluxes in a forest voxel grid.
}
