% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM_LW.R
\name{lw_two_stream}
\alias{lw_two_stream}
\title{Two-stream RTM for longwave radiation and for a single column existing of layers/voxels with density values (\%).
This model is inspired by the longwave RTM of the Ecosystem Demography model (ED 2.2).}
\usage{
lw_two_stream(
  soil_reflect,
  density,
  F_sky_lw,
  temperature,
  T_soil,
  T_macro,
  Kd,
  omega,
  beta
)
}
\arguments{
\item{soil_reflect}{Soil albedo (\%)}

\item{density}{Density of each layer (\%)}

\item{F_sky_lw}{Longwave radiation at top canopy (W/m2)}

\item{temperature}{Forest surface temperature of each layer (K)}

\item{T_soil}{Soil temperature at 8 cm depth (K)}

\item{T_macro}{Macrotemperature outside the forest (K)}

\item{Kd}{Longwave extinction coefficient, per unit density}

\item{omega}{Longwave scattering coefficient}

\item{beta}{Fraction of scattering in the backward direction for longwave radiation (\%)}
}
\value{
Dataframe containing the combined results of the 2D longwave RTM
}
\description{
Two-stream RTM for longwave radiation and for a single column existing of layers/voxels with density values (\%).
This model is inspired by the longwave RTM of the Ecosystem Demography model (ED 2.2).
}
