% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_temperature_fluxes.R
\name{plot_rad}
\alias{plot_rad}
\title{Function to create radiation flux plot at 1m height}
\usage{
plot_rad(micro_grid, down_radiation, output_path)
}
\arguments{
\item{micro_grid}{Grid with micro variables}

\item{output_path}{The output_path to store the plots}

\item{net_radiaiton}{Vector with net radiation}
}
\value{
flux plot at 1m height of model output vs PAR sensors
}
\description{
Function to create radiation flux plot at 1m height
}
