% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots}
\alias{plots}
\title{Function to create several temperature plots}
\usage{
plots(micro_grid, T_air_vec)
}
\arguments{
\item{micro_grid}{Grid with micro variables}

\item{T_air_vec}{Vector with air temperatures}
}
\value{
Several temperature plots (plotted and saved)
}
\description{
Function to create several temperature plots
}
