% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_temperature_fluxes.R
\name{plots_flux}
\alias{plots_flux}
\title{Function to create several flux plots}
\usage{
plots_flux(
  micro_grid,
  net_radiation,
  sensible_flux,
  latent_flux,
  G,
  output_path,
  datetime
)
}
\arguments{
\item{micro_grid}{Grid with micro variables}

\item{sensible_flux}{Vector with sensible heat fluxes}

\item{latent_flux}{Vector with latent heat fluxes}

\item{G}{Vector with ground heat fluxes}

\item{output_path}{The output_path to store the plots}

\item{datetime}{yyyy-mm-dd hh:mm:ss in UTC}

\item{net_radiaiton}{Vector with net radiation}
}
\value{
Several flux plots (plotted and saved)
}
\description{
Function to create several flux plots
}
