% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_longwave_radiation.R
\name{plots_lw}
\alias{plots_lw}
\title{Function to create several longwave radiation plots}
\usage{
plots_lw(lw_rad_2D)
}
\arguments{
\item{sw_rad_2D}{Dataframe with shortwave radiative values}
}
\value{
Several radiation plots (plotted and saved)
}
\description{
Function to create several longwave radiation plots
}
