% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_temperature_fluxes.R
\name{plots_temp}
\alias{plots_temp}
\title{Function to create several temperature plots}
\usage{
plots_temp(micro_grid, T_air_vec, output_path, datetime)
}
\arguments{
\item{micro_grid}{Grid with micro variables}

\item{T_air_vec}{Vector with air temperatures}

\item{output_path}{The output_path to store the plots}

\item{datetime}{yyyy-mm-dd hh:mm:ss in UTC}
}
\value{
Several temperature plots (plotted and saved)
}
\description{
Function to create several temperature plots
}
