% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_temperature_fluxes.R
\name{plots_temp_flux}
\alias{plots_temp_flux}
\title{Function to create several temperature plots}
\usage{
plots_temp_flux(
  micro_grid,
  T_air_vec,
  net_radiation,
  sensible_flux,
  latent_flux,
  G
)
}
\arguments{
\item{micro_grid}{Grid with micro variables}

\item{T_air_vec}{Vector with air temperatures}

\item{sensible_flux}{Vector with sensible heat fluxes}

\item{latent_flux}{Vector with latent heat fluxes}

\item{G}{Vector with ground heat fluxes}

\item{net_radiaiton}{Vector with net radiation}
}
\value{
Several temperature plots (plotted and saved)
}
\description{
Function to create several temperature plots
}
