% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ForEdgeClim.R
\name{run_foredgeclim}
\alias{run_foredgeclim}
\title{Execute the microclimate model ForEdgeClim}
\usage{
run_foredgeclim(structure_grid, datetime, trace_idx = NULL)
}
\arguments{
\item{structure_grid}{Dataframe with structural density values}

\item{datetime}{Datetime object representing the current simulation time}
}
\value{
Dataframe with simulated microclimate temperatures and fluxes
}
\description{
Execute the microclimate model ForEdgeClim
}
