% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{saturated_vapor_pressure}
\alias{saturated_vapor_pressure}
\title{Function for saturated vapor pressure; this is the empirical equation of Tetens}
\usage{
saturated_vapor_pressure(temp)
}
\arguments{
\item{temp}{Surface temperature}
}
\value{
Saturated vapor pressure
}
\description{
Function for saturated vapor pressure; this is the empirical equation of Tetens
}
