% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM_SW.R
\name{shortwave_two_stream_RTM}
\alias{shortwave_two_stream_RTM}
\title{2D (vertical & horizontal) shortwave two-stream RTM}
\usage{
shortwave_two_stream_RTM(
  datetime,
  lat,
  lon,
  voxel_grid,
  F_sky_diff_init,
  F_sky_dir_init,
  omega_g_v,
  omega_g_h,
  Kd_v,
  Kb_v,
  Kd_h,
  Kb_h,
  omega,
  beta,
  beta0
)
}
\arguments{
\item{datetime}{Datetime object representing the current simulation time}

\item{lat}{Latitude of the location}

\item{lon}{Longitude of the location}

\item{voxel_grid}{Dataframe containing the voxel grid with density values}

\item{F_sky_diff_init}{Initial diffuse shortwave radiation (W/m2)}

\item{F_sky_dir_init}{Initial direct shortwave radiation (W/m2)}

\item{omega_g_v}{Ground reflectance for vertical RTM}

\item{omega_g_h}{Ground reflectance for horizontal RTM}

\item{Kd_v}{Diffuse extinction coefficient for vertical RTM}

\item{Kb_v}{Direct extinction coefficient for vertical RTM}

\item{Kd_h}{Diffuse extinction coefficient for horizontal RTM}

\item{Kb_h}{Direct extinction coefficient for horizontal RTM}

\item{omega}{Scattering coefficient}

\item{beta}{Fraction of scattered diffuse radiation in the backward direction}

\item{beta0}{Fraction of scattered direct radiation in the backward direction}
}
\value{
Dataframe containing the combined results of the 2D shortwave RTM
}
\description{
2D (vertical & horizontal) shortwave two-stream RTM
}
