% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_temperatures.R
\name{sin_lag}
\alias{sin_lag}
\title{Sinus for near-surface air temperature: time lagged wrt macro air temperature}
\usage{
sin_lag(datetime)
}
\arguments{
\item{datetime}{Datetime object representing the current simulation time}
}
\value{
TG_diff Lagging temperature correction to determine soil temperature
}
\description{
Sinus for near-surface air temperature: time lagged wrt macro air temperature
}
