% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTM_SW.R
\name{sw_two_stream}
\alias{sw_two_stream}
\title{Two-stream RTM for shortwave radiation and for a single column existing of layers/voxels with density values (\%).
This model is inspired by the shortwave RTM of the Ecosystem Demography model (ED 2.2).}
\usage{
sw_two_stream(
  soil_reflect,
  density,
  F_sky_diff,
  F_sky_dir,
  Kd,
  Kb,
  omega,
  beta,
  beta0
)
}
\arguments{
\item{soil_reflect}{Soil albedo (\%)}

\item{density}{Density of each layer (\%)}

\item{F_sky_diff}{Diffuse radiation (W/m2)}

\item{F_sky_dir}{Direct beam radiation (W/m2)}

\item{Kd}{Diffuse extinction coefficient, per unit density}

\item{Kb}{Direct beam extinction coefficient, per unit density}

\item{omega}{Shortwave scattering coefficient}

\item{beta}{Fraction of scattering in the backward direction for diffuse radiation (\%)}

\item{beta0}{Fraction of scattering in the backward direction for direct beam radiation (\%)}
}
\value{
Dataframe containing the combined results of the 2D shortwave RTM
}
\description{
Two-stream RTM for shortwave radiation and for a single column existing of layers/voxels with density values (\%).
This model is inspired by the shortwave RTM of the Ecosystem Demography model (ED 2.2).
}
