% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{calculate_H}
\alias{calculate_H}
\title{Function to calculate sensible heat transfer between forest surface and air}
\usage{
calculate_H(T_surf, T_air)
}
\arguments{
\item{T_surf}{Forest surface temperature}

\item{T_air}{Air temperature}
}
\value{
H Sensible heat flux
}
\description{
Function to calculate sensible heat transfer between forest surface and air
}
