% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_input_variables_and_params.R
\name{create_input_data}
\alias{create_input_data}
\title{Generate standard input data for ForEdgeClim}
\usage{
create_input_data(summer_day = TRUE)
}
\arguments{
\item{summer_day}{Logical value: TRUE for day, FALSE for night}
}
\value{
global variables and parameters
}
\description{
Generate standard input data for ForEdgeClim
}
