% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_balance.R
\name{de_dT}
\alias{de_dT}
\title{Function to calculate the derivative of the energy balance closure error to surface temperature}
\usage{
de_dT(temp)
}
\arguments{
\item{temp}{Surface temperature}
}
\value{
Derivative of energy balance wrt surface temperature
}
\description{
Function to calculate the derivative of the energy balance closure error to surface temperature
}
