% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid_and_DTM_from_TLS.R
\name{generate_DTM_grid_TLS}
\alias{generate_DTM_grid_TLS}
\title{Generate a voxelized point density grid from a LAS file}
\usage{
generate_DTM_grid_TLS(las_file, voxel_size = 1)
}
\arguments{
\item{las_file}{Path to the LAS file}

\item{voxel_size}{Numeric value defining the voxel size (default = 1m).}
}
\value{
A data.table with columns: X, Y, Z, and density.
}
\description{
This function reads a LAS file and creates a 3D voxel grid where each voxel represents
a 1x1x1m^3 volume (or a user-defined size). The density of each voxel is computed
based on the number of points in the point cloud. The output is a data.table containing
rescaled X, Y, and Z coordinates and a normalized density value.
}
\details{
Additionally, a 3D plot of the voxel grid and a digital terrain model (DTM) are generated and saved as PNG files in
the "Output" directory, i.e. 'output_path' as defined by the user (see script 'testing_foredgeclim.R')
}
