% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_RMI_observations}
\alias{import_RMI_observations}
\title{Function to import RMI national weather station observations (macro
temperature and incoming longwave radiation)}
\usage{
import_RMI_observations(datetime)
}
\value{
macro temperature and downward from sky longwave radiation
}
\description{
Function to import RMI national weather station observations (macro
temperature and incoming longwave radiation)
}
