% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_observations.R
\name{import_soil_temperature}
\alias{import_soil_temperature}
\title{Function to import TOMST observations (soil temperature at tower and
observations along the transect)}
\usage{
import_soil_temperature(datetime)
}
\value{
T_soil_deep The soil temperature at 6cm deep at the position of the tower
}
\description{
Function to import TOMST observations (soil temperature at tower and
observations along the transect)
}
