% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_DTM_structure.R
\name{plots_dtm_struct}
\alias{plots_dtm_struct}
\title{Function to create plot of DTM and density structure}
\usage{
plots_dtm_struct(dtm, grid, output_path)
}
\arguments{
\item{dtm}{Dataframe with digital terrain model}

\item{grid}{Dataframe with voxelized density values}

\item{output_path}{The output_path to store the plots}
}
\value{
Several radiation plots (plotted and saved)
}
\description{
Function to create plot of DTM and density structure
}
