% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_shortwave_radiation.R
\name{plots_sw}
\alias{plots_sw}
\title{Function to create several shortwave radiation plots}
\usage{
plots_sw(sw_rad_2D, output_path)
}
\arguments{
\item{sw_rad_2D}{Dataframe with shortwave radiative values}

\item{output_path}{The output_path to store the plots}
}
\value{
Several radiation plots (plotted and saved)
}
\description{
Function to create several shortwave radiation plots
}
