% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ForEdgeClim.R
\name{run_foredgeclim}
\alias{run_foredgeclim}
\title{Execute the microclimate model ForEdgeClim}
\usage{
run_foredgeclim()
}
\arguments{
\item{input_data}{Dataframe with input variables}
}
\value{
Dataframe with simulated microclimate temperatures and fluxes
}
\description{
Execute the microclimate model ForEdgeClim
}
